Reimlementation of RowMajor, and temporary removal of ColMajor.
This time, there is a single implementation of ColMajor for static,
dynamic, normal and sliced matrices.

Currently only static works, and test/mat_test2.cc compiles and produces
the correct output.

The general principle behind this version is to pass around metadata at run
time in cases it is needed, and then ignore it if it is not.