Rewritten the slicing interface.
Now, matrix can pass on a generic SliceSpec object which will hold data needed
to perform a slice. This is forwarded by the Base class on to GenericMBase.
However, in GenericMBase, a mechanism is used to block the system to silently
ignore superfluous data.

There is no public interface yet.

test/test_foreign.cc  demos the private interface. Exacly one class in SliceSpec
will allow the code to compile.