Added some debugging macros to initialize data to values likely to cause failure.
TOON_INITIALIZE_NAN  initialize all data to NaN

TOON_INITIALIZE_VAL  initialize all data to the expansion of this macro

TOON_INITIALIZE_RANDOM initialize all data to random bytes. This will not
compile if you make a Vector of a non-POD type. This is seeded off
std::time, so the results will be nicely non-reproducable to a granularity
of 1 second. A very cheap and nasty random number generator has been used.