Added a "pure" unscaled identity object, which carries no data.
This is VERY important, since there is a global identity object. Previously
the global Identity carried data and so had to be constructed. This caused
nasty bugs due to the unspecified order of initialization of global objects.

The bug can be observed with the following:
g++  identity_test.cc -fmudflap -lmudflap -g -ggdb
MUDFLAP_OPTIONS=-viol-gdb ./a.out