Special new "One" type which represents the value 1 with no data. This is used for global objects, so that their constructors can be trivial.
Special hacking of Field required. One is not a field by itself, so
IsField<One>::value is 0, but it can be combined with a field
arithmetically and the result is a field, so Field<One,T> and Field<T,One>
is defined.

One can cast itself to any type, and be added, subtracted or multiplied.

Also, special cases for negation are required since -One is not representable
using One.