Add compatibility with CLAPACK.
In normal FORTRAN, int maps to Integer (on 32 and 64 bit systems).

In CLAPACK, long maps to Integer. Note that CLAPACK cannot be used out of
the box with TooN as it requires libf2c which defines main().

To set the FORTRAN integer type use -DTOON_CLAPACK (to make it long int)  or
-DTOON_FORTRAN_INTEGER=long to set it to an arbitrary type.