Added cost estimation and optional caching.
Call the .perhaps_cache(int) method with the number of evaluations
of each element to be made (e.g. number of rows in a matrix* vector
multiply).

Looking at the generated asm, the optional caching doesn't appear to
incur any run-time overhead if the .perhaps_cache() method is not
called.